requirements.txt
psycopg2
pySmartDL==1.3.3
requests==2.24.0
git+https://github.com/pokurt/pytube3.git
git+https://github.com/pyrogram/pyrogram
TgCrypto
pydrive==1.3.1
sqlalchemy
psycopg2-binary
bs4==0.0.1
simplejson
removebg==0.4
telegraph==1.4.1
aiohttp==3.6.2
pyDownload==0.0.1.dev4
pillow==7.1.2
gtts==2.1.1
gitpython
pafy==0.5.5
git+https://github.com/ytdl-org/youtube-dl
gpytranslate
currencyconverter
uptime
py-cpuinfo
svglib
reportlab
aiofiles
speedtest-cli
timeago
httplib2==0.15.0
coffeehouse==2.2.2
pycurl==7.43.0.5
tracemoepy
emoji==0.5.4
beautifulsoup4==4.9.1
pathlib==1.0.1
typing==3.7.4.1
spamwatch==0.3.0
htmlement==1.0.0
pyYAML
mega.py==1.0.8
twistdl